+++ canvas_sdk/v1/data/__init__.py
@@ -18,12 +18,14 @@ from .claim import (
     ClaimComment,
     ClaimCoverage,
     ClaimLabel,
+    ClaimMetadata,
     ClaimPatient,
     ClaimProvider,
     ClaimQueue,
     ClaimSubmission,
     InstallmentPlan,
 )
+from .claim_banner_alert import BannerAlertIntent, BannerAlertStatus, ClaimBannerAlert
 from .claim_diagnosis_code import ClaimDiagnosisCode
 from .claim_line_item import ClaimLineItem, ClaimLineItemDiagnosisCode
 from .command import Command
@@ -181,11 +183,15 @@ __all__ = __exports__ = (
     "CareTeamRole",
     "ChargeDescriptionMaster",
     "Claim",
+    "ClaimBannerAlert",
+    "BannerAlertStatus",
+    "BannerAlertIntent",
     "ClaimComment",
     "ClaimCoverage",
     "ClaimDiagnosisCode",
     "ClaimLabel",
     "ClaimLineItem",
+    "ClaimMetadata",
     "ClaimLineItemDiagnosisCode",
     "ClaimPatient",
     "ClaimProvider",