+++ CHANGELOG.md
@@ -1,6 +1,30 @@
 # CHANGELOG
 
 
+## v0.105.0 (2026-02-26)
+
+### Features
+
+- Create claim metadata model and effect
+  ([#1514](https://github.com/canvas-medical/canvas-plugins/pull/1514),
+  [`fcf3f62`](https://github.com/canvas-medical/canvas-plugins/commit/fcf3f622ca70adf2015af9d12eaa18565048553a))
+
+
+## v0.104.0 (2026-02-24)
+
+### Documentation
+
+- Bulk update references to BaseProtocol to prefer BaseHandler
+  ([#1519](https://github.com/canvas-medical/canvas-plugins/pull/1519),
+  [`1214fa2`](https://github.com/canvas-medical/canvas-plugins/commit/1214fa2768f37d346f0f60c608f76ff3601c4d2b))
+
+### Features
+
+- Enable ClaimBannerAlert effects
+  ([#1485](https://github.com/canvas-medical/canvas-plugins/pull/1485),
+  [`cb13d4a`](https://github.com/canvas-medical/canvas-plugins/commit/cb13d4a803221d10ee8f8c43bf77fbb4b3f1d2a7))
+
+
 ## v0.103.1 (2026-02-23)
 
 ### Bug Fixes